// backend/src/interfaces/DetallePedido.ts
//Una interfaz en TypeScript es una forma de definir la estructura de un objeto. 
//Es muy util para definir tipos y validar que los datos cumplan con ciertos requisitos.

export interface DetallePedido {
    idproducto: number;
    cantidad: number;
    subtotal: number; // Agregar esta propiedad
}